+++ src/main/java/com/example/oatnote/memotag/MemoTagService.java
@@ -39,7 +39,7 @@ import com.example.oatnote.memotag.service.memoTagRelation.MemoTagRelationServic
 import com.example.oatnote.memotag.service.tag.TagService;
 import com.example.oatnote.memotag.service.tag.edge.model.TagEdge;
 import com.example.oatnote.memotag.service.tag.model.Tag;
-import com.example.oatnote.web.models.Criteria;
+import com.example.oatnote.web.model.Criteria;
 
 import lombok.RequiredArgsConstructor;
