+++ src/main/java/com/example/oatnote/memotag/dto/ChildTagsWithMemosResponse.java
@@ -6,8 +6,7 @@ import java.util.List;
 
 import org.springframework.data.domain.Page;
 
-import com.example.oatnote.memotag.dto.innerDto.TagResponse;
-import com.example.oatnote.web.models.Criteria;
+import com.example.oatnote.web.model.Criteria;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
